+++ lockfile_generator/src/gradle.rs
@@ -1,12 +1,25 @@
 //! Java gradle ecosystem.
 
 use std::ffi::OsStr;
+use std::fs;
+use std::io::Write;
 use std::path::{Path, PathBuf};
-use std::process::Command;
+use std::process::{Command, Stdio};
 
-use crate::{Error, Generator, Result};
+use tempfile::NamedTempFile;
+
+use crate::{Error, FileRelocator, Generator, Result};
 
 pub struct Gradle;
+
+/// Init script passed to gradle to force writing lockfiles.
+const INIT_SCRIPT: &[u8] = b"
+allprojects {
+    dependencyLocking {
+        lockAllConfigurations()
+    }
+}
+";
 
 impl Generator for Gradle {
     fn lockfile_path(&self, manifest_path: &Path) -> Result<PathBuf> {
@@ -17,9 +30,8 @@ impl Generator for Gradle {
     }
 
     fn command(&self, _manifest_path: &Path) -> Command {
-        let mut command = Command::new("gradle");
-        command.args(["dependencies", "--write-locks"]);
-        command
+        // NOTE: We use a custom command to pass the init script.
+        unreachable!()
     }
 
     fn tool(&self) -> &'static str {
@@ -34,5 +46,58 @@ impl Generator for Gradle {
         } else {
             Err(Error::InvalidManifest(manifest_path.to_path_buf()))
         }
+    }
+
+    fn generate_lockfile(&self, manifest_path: &Path) -> Result<String> {
+        self.check_prerequisites(manifest_path)?;
+
+        let canonicalized = fs::canonicalize(manifest_path)?;
+        let project_path = canonicalized
+            .parent()
+            .ok_or_else(|| Error::InvalidManifest(manifest_path.to_path_buf()))?;
+
+        // Move files which interfere with lockfile generation.
+        let _relocators = self
+            .conflicting_files(&canonicalized)?
+            .drain(..)
+            .map(FileRelocator::new)
+            .collect::<Result<Vec<_>>>()?;
+
+        // Create temporary init script file.
+        let mut init_file = NamedTempFile::new()?;
+        init_file.write_all(INIT_SCRIPT)?;
+        let init_path = init_file.path().to_string_lossy();
+
+        // Generate lockfile at the target location.
+        let mut command = Command::new("gradle");
+        command.args(["dependencies", "--init-script", &init_path, "--write-locks"]);
+        command.current_dir(project_path);
+        command.stdin(Stdio::null());
+        command.stdout(Stdio::null());
+
+        // Provide better error message, including the failed program's name.
+        let output = command.output().map_err(|err| {
+            let program = format!("{:?}", command.get_program());
+            Error::ProcessCreation(program, self.tool().to_string(), err)
+        })?;
+
+        // Ensure generation was successful.
+        if !output.status.success() {
+            return Err(Error::NonZeroExit(output));
+        }
+
+        // Ensure lockfile was created.
+        let lockfile_path = self.lockfile_path(&canonicalized)?;
+        if !lockfile_path.exists() {
+            return Err(Error::NoLockfileGenerated);
+        }
+
+        // Read lockfile contents.
+        let lockfile = fs::read_to_string(&lockfile_path)?;
+
+        // Cleanup lockfile.
+        fs::remove_file(lockfile_path)?;
+
+        Ok(lockfile)
     }
 }