+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed
+
+- Gradle not generating lockfiles without `dependencyLocking` in the manifest
+
 ## 6.6.1 - 2024-06-18
 
 ### Fixed